+++ Touhou Spell Up!!!/Assets/Scripts/DanmakuGame/Patterns/Move/MovePatternBase.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+using Cysharp.Threading.Tasks;
+using System.Threading;
+
+public abstract class MovePatternBase : PatternBase
+{
+    public override UniTask ExecuteImpl(Mover mover, Shooter _, Bullet __, CancellationToken token)
+    {
+        // 実際の処理はさらにサブクラスのExecuteMoveに委譲する
+        return ExecuteMove(mover, token);
+    }
+
+    // サブクラスが実装するための新しい抽象メソッド
+    public abstract UniTask ExecuteMove(Mover mover, CancellationToken token);
+}